% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pSup.R
\name{calculate_pSup}
\alias{calculate_pSup}
\title{Calculate pSup}
\usage{
calculate_pSup(
  wide_data,
  chains = 4,
  iter = 1000,
  control = list(adapt_delta = 0.85)
)
}
\arguments{
\item{wide_data}{A wider data frame}

\item{chains}{A number, default to 4}

\item{iter}{A number, default to 1000}

\item{control}{A list of parameters, default to list(adapt_delta = 0.85)}
}
\description{
Calculate the proportion of each transcript in the Supernatant.
}
\details{
The input data should be the wide format in terms of fractions.
Function get_paras_median from this package is called in this function. Output
a data frame that includes a variable pSup calculated for each transcript.
}
\examples{
 
\dontrun{
calculate_pSup(wide_data)
calculate_pSup(wide_data = mydata)
}
}
\seealso{
[rstan::sampling()], `browseVignettes("rstan")`
}
\keyword{pSup}
\keyword{pellet}
\keyword{proportion}
\keyword{supernatant}
