% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pSup.R
\name{each_mRNA_pSup}
\alias{each_mRNA_pSup}
\title{Represent pSup for each transcript}
\usage{
each_mRNA_pSup(
  wide_data,
  chains = 4,
  iter = 1000,
  control = list(adapt_delta = 0.85)
)
}
\arguments{
\item{wide_data}{A wider data frame}

\item{chains}{A number, default to 4}

\item{iter}{A number, default to 1000}

\item{control}{A list of parameters, default to list(adapt_delta = 0.85)}
}
\description{
Clear format to represent the proportion of each transcript in the Supernatant.
}
\details{
The input data should be the wide format in terms of fractions.
Function calculate_pSup from this package is called in this function. Output 
a data frame that shows the pSup value of each transcript in each fraction.
}
\examples{
 
\dontrun{
each_mRNA_pSup(wide_data)
each_mRNA_pSup(wide_data = mydata)
}
}
\seealso{
[rstan::sampling()], `browseVignettes("rstan")`
}
\keyword{pSup}
\keyword{supernatant}
