% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{get_para_sta}
\alias{get_para_sta}
\title{parameters statistical results}
\usage{
get_para_sta(
  wide_data,
  chains = 4,
  iter = 1000,
  control = list(adapt_delta = 0.85),
  ...
)
}
\arguments{
\item{wide_data}{A wider data frame}

\item{chains}{A number, default to 4}

\item{iter}{A number, default to 1000}

\item{control}{A list of parameters, default to list(adapt_delta = 0.85)}

\item{...}{Other arguments passed to `rstan::sampling` (e.g. warmup, seed).}
}
\value{
A data frame
}
\description{
Get the estimated statistical results of parameters
}
\details{
The input data should be the wide format in terms of fractions.
Function model_fit from this package is called in this function.
Output a data frame that includes a statistical result of parameters.
}
\examples{
get_para_sta(wide_data, chains = 1, iter = 500)
\dontrun{
get_para_sta(wide_data)
get_para_sta(wide_data, chains = 4)
get_para_sta(wide_data, iter = 1500, control = list(adapt_delta = 0.85))
get_para_sta(wide_data, control = list(adapt_delta = 0.85))
}
}
\seealso{
[rstan::sampling()], `browseVignettes("rstan")`
}
\keyword{estimation}
\keyword{parameter}
