% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{get_wide_Fraction}
\alias{get_wide_Fraction}
\title{Change the data format}
\usage{
get_wide_Fraction(dir_in, file)
}
\arguments{
\item{dir_in}{The input directory name}

\item{file}{The input Samplesheet file name}
}
\description{
Covert the long Fraction coloum to the wide format.
}
\details{
In this function, function read_count_files from this package is called.
Output a wider data frame for representing the count value of each trancript
for each fraction.
}
\examples{
\dontrun{
get_wide_Fraction("inst/extdata","Samplesheet")
get_wide_Fraction(dir_in = "inst/extdata", file = "Samplesheet.txt")
}
## Select the columns that we need
## Generate a wider data format to get the variables of fractions
}
\keyword{counts}
\keyword{each}
\keyword{fraction}
