% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Model fit}
\usage{
model_fit(
  wide_data,
  chains = 4,
  iter = 1000,
  control = list(adapt_delta = 0.85),
  ...
)
}
\arguments{
\item{wide_data}{A wider data frame}

\item{chains}{A number, defaulting to 4}

\item{iter}{A number, defaulting to 1000}

\item{control}{A list of parameters, defaulting to list(adapt_delta = 0.85)}

\item{...}{Other arguments passed to `rstan::sampling` (e.g. warmup, seed).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Fit the bayesian statistical model to the counts data.
}
\details{
The input data should be the wide format in terms of fractions.
It will be coverted to a list of vectors through the function compose_data from 
package tidybayes for model fitting. Compile_model is the pre_loaded data that 
stores the compiled bayesian statistical model.
}
\examples{
 
model_fit(wide_data, chains = 1, iter = 500)
\dontrun{
model_fit(wide_data)
model_fit(wide_data, chains = 4)
model_fit(wide_data = mydata, iter = 1500, control = list(adapt_delta = 0.85))
}

}
\seealso{
[rstan::sampling()], `browseVignettes("rstan")`
}
\keyword{fit}
\keyword{model}
