% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Read Samplesheet}
\usage{
read_samplesheet(dir_in, file)
}
\arguments{
\item{dir_in}{The input directory name}

\item{file}{The input Samplesheet file name}
}
\value{
A single data frame
}
\description{
Read the Samplesheet file that is provided by users.
}
\details{
This function is to read the samplesheet file that is provided by users.
Useres need to make sure that the input directory is under the working directory 
and the samplesheet file is in the input directory.
}
\examples{
 
\dontrun{
read_samplesheet("inst/extdata","Samplesheet")
read_samplesheet(dir_in = "inst/extdata",file = "Samplesheet")
}
samplesheet_example
}
\keyword{samplesheet}
