% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wide_data}
\alias{wide_data}
\title{wide_data}
\format{
A data frame.
\describe{
  \item{wide_data}{A wide-format data frame that clearly displays all the count values of 20 different transcripts in different fractions.
}
}
}
\source{
\href{https://github.com/thebestecho/RNAFracQuant/tree/Msc_project}{RNAFracQuant}
}
\usage{
wide_data
}
\description{
A single data frame that is in wide format. This data is used as the input data 
in vignettes and tests of our package RNAFracQuant.
}
\keyword{wide_data}
