% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pSup.R
\name{write_results}
\alias{write_results}
\title{Write results into files}
\usage{
write_results(...)
}
\arguments{
\item{...}{Unlimited arguments, which are supposed to be data frames.}
}
\description{
Write results into files
}
\details{
This function takes the outputs of functions in RNAFracQuant as the input arguments 
and then writes them into text files under the same "Results" directory.
}
\examples{
 
\dontrun{
write_results(paras_median, data_pSup)
write_results(data1, data2, data3, data4)
}
}
\seealso{
[readr::write_tsv()], `browseVignettes("readr")`
}
\keyword{results}
